+++ changelog.md
@@ -337,7 +337,10 @@
 
 - Added setCurrentException for JS backend.
 
+- Added `tables.getPtr`.
+
 - Added `dom.scrollIntoView` proc with options
+
 
 ## Language changes